Factor out function-related codes
